requirements.txt
pyserial>=3.5
numpy>=1.24
scipy>=1.11
matplotlib>=3.7
pyautogui>=0.9.54
requests>=2.31
flask>=3.0
